Reorganized assets directory to have an img subdir and fixed all imports broken by that change.
