+++ src/assets/img/dragon-outline.png
