+++ src/assets/img/third-place-dragon.png
