+++ src/assets/img/voting-booth.png
