+++ src/components/pages/LandingPage/DragonBoi/DragonBoi.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import dragonBoi from '../../../../assets/dragon-boi.svg';
+import dragonBoi from '../../../../assets/img/dragon-boi.svg';
 
 const DragonBoi = (): React.ReactElement => {
   return (
