+++ src/components/pages/ResultsPage/RenderResultsPage.tsx
@@ -1,12 +1,9 @@
 import React from 'react';
-
 import { useRecoilValue } from 'recoil';
+import celebration from '../../../assets/img/celebration-station.png';
+import dragonBoi from '../../../assets/img/dragon-boi.svg';
 import { results } from '../../../state';
-
 import { Header, Histogram, SubCard, ThoughtBubble } from '../../common';
-
-import dragonBoi from '../../../assets/dragon-boi.svg';
-import celebration from '../../../assets/celebration-station.png';
 
 const RenderResultsPage = (): React.ReactElement => {
   const winner = useRecoilValue(results.winner);
